Conflicts:
	contracts/utils/cryptography/ECDSA.sol
	contracts/utils/cryptography/SignatureChecker.sol

+++ contracts/utils/Multicall.sol
@@ -11,11 +11,11 @@ import "./Address.sol";
  */
 abstract contract Multicall {
     /**
-    * @dev Receives and executes a batch of function calls on this contract.
-    */
+     * @dev Receives and executes a batch of function calls on this contract.
+     */
     function multicall(bytes[] calldata data) external returns (bytes[] memory results) {
         results = new bytes[](data.length);
-        for (uint i = 0; i < data.length; i++) {
+        for (uint256 i = 0; i < data.length; i++) {
             results[i] = Address.functionDelegateCall(address(this), data[i]);
         }
         return results;
